refactored file structure,breakdown large components
